+++ ClosestStringProblem/aco.hpp
@@ -28,7 +28,7 @@ class ACO {
     void construct(Ant *current_ant);
     void local_search(Ant * ant);
     void local_search2(Ant * ant);
-    double heuristic_information(Ant *current_ant, int idx, int char_idx);
+    double heuristic_information(Ant *current_ant, long int idx, long int char_idx);
     void update_pheromone_trails(Ant *global_best, double tau_min, double tau_max);
 
 public:
